+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Send Events.SESSION_READY when the main window is created to initialize widget components
+
+### Changed
+
+- Session.change receive a position argument instead of a keyword one
+
 ## 0.14.0
 
 ### Added