+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Move Alarm plugin code to tomate-gtk repository
+- Move notify plugin to tomate-gtk repository
+
+## 0.21.0
+
+### Changed
+
+- Move alarm plugin to tomate-gtk repository
 
 ## 0.20.0